+++ packages/store/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/store
 
+## 1.3.2
+
+### Patch Changes
+
+-   cc65eec: adds new vscode sap systems extension
+
 ## 1.3.1
 
 ### Patch Changes