+++ packages/ui-service-inquirer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/ui-service-inquirer
 
+## 0.2.43
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.6.42
+-   @sap-ux/odata-service-inquirer@2.9.14
+-   @sap-ux/telemetry@0.6.36
+-   @sap-ux/inquirer-common@0.8.10
+
 ## 0.2.42
 
 ### Patch Changes